Chỉnh sửa SnakeKun
1.Sửa SnakeKun(map[Height][Width]) thành createSnake();
2.Thêm vào mảng 2 chiều MapData::_map[MAX][MAX] để lưu tọa độ của snake.
3.Thêm vào hàm countFood để đếm điểm.
4.Hoàn thiện hàm update()
