+++ ConSnake/Menu.h
@@ -1,0 +1,13 @@
+#pragma once
+#include "Library.h"
+
+class Menu
+{
+public:
+    Menu();
+    ~Menu();
+    void StartMenu();
+    GameMode GamemodeMenu();
+};
+
+
