+++ ConSnake/SnakeKun.cpp
@@ -1,10 +1,10 @@
 #include "SnakeKun.h"
 
-SnakeKun::SnakeKun(MapData map[WIDTH][HEIGHT])
+void SnakeKun::createSnake()
 {
     _length = INIT_SNAKE_LENGTH;
     _body[0] = { WIDTH / 2,HEIGHT / 2 };
-    _direction = INPUT::userInput;
+    _direction = Input::userInput;
     _foodCounter = INIT_FOOD_COUNTER;
     for (int i = 1; i < _length; i++)
     {
@@ -12,21 +12,89 @@ SnakeKun::SnakeKun(MapData map[WIDTH][HEIGHT])
         _body[i].y = _body[i - 1].y - dy[(int)_direction];
     }
     for (int i = 0; i < _length; i++)
-        map[_body[i].y][_body[i].x] = MapData::SNAKE;
+        _map[_body[i].y][_body[i].x] = MapData::SNAKE;
 }
 
-void SnakeKun::update()
+void SnakeKun::update(unsigned short int& delay)
 {
     //dia chi tam thoi cua ran
+    int i;
+    int delay_slow = delay;
     Coordinate _prev[WIDTH * HEIGHT];
 
-    for (int i = 0; i < _length; i++)
+    for (i = 0; i < _length; i++)
     {
         _prev[i].x = _body[i].x;     
         _prev[i].y = _body[i].y;
     }
-    if (INPUT::userInput != EXIT && !oppositeDirection(_direction, INPUT::userInput))
-       
+    if (Input::userInput != SnakeDirection::EXIT && !oppositeDirection(_direction, Input::userInput))
+        _direction = Input::userInput;
+
+    //di chuyen dau con snake theo huong cua input (2)
+    _body[0].x = _prev[0].x + dx[(int)_direction];		
+    _body[0].y = _prev[0].y + dy[(int)_direction];
+
+    //neu tu dam vao minh thi gameover(3)
+    if (_map[_body[0].y][_body[0].x] < MapData::NOTHING)
+    {
+        item = -1;     
+        return;
+    }
+
+    //neu an duoc thi cong diem(4)
+    if (_map[_body[0].y][_body[0].x] == FOOD)
+    {
+        countFood(delay);
+    }
+    else
+    {
+        _map[_body[_length - 1].y][_body[_length - 1].x] = MapData::NOTHING;
+        item = (short int)MapData::NOTHING;
+        gotoXY(_body[_length - 1].x, _body[_length - 1].y);		// neu snake chua tang kich co thi(6)
+        cout << " ";						                    // xoa phan cuoi cua snake
+    }
+
+    for (i = 1; i < _length; i++)
+    {
+        _body[i].x = _prev[i - 1].x;	//snake di chuyen theo huong ban dau(7)
+        _body[i].y = _prev[i - 1].y;
+    }
+
+    //dua du lieu snake vao map(8)
+    for (i = 0; i < _length; i++)
+        _map[_body[i].y][_body[i].x] = MapData::SNAKE;
+
+    //neu ran di len/xuong thi giam toc do(9)
+    if (Input::userInput == SnakeDirection::UP || Input::userInput == SnakeDirection::DOWN)
+    {
+        delay_slow += (delay * 25) / 100;
+        Sleep(delay_slow);
+    }
+    else Sleep(delay);
+
+}
+
+void SnakeKun::countFood(unsigned short int& delay)
+{
+    //tinh diem khac nhau, tuy gamemode (5)
+    switch (_gameMode) 
+    {
+    case GameMode::EASY:
+        _foodCounter += (_length * 5); break;
+    case GameMode::NORMAL:
+        _foodCounter += (_length * 10); break;
+    case GameMode::HARD:
+        _foodCounter += (_length * 15); break;
+    case GameMode::SPECIAL:
+        _foodCounter += (_length * (_length - 3)); break;
+    }
+    // tang kich thuoc ran khi ran an(6)
+    _length++; 
+    item = FOOD;
+    if (_gameMode == GameMode::SPECIAL && _length <= 104) {
+        //tang toc do trong gamemode special
+        --delay; 
+    }
 }
 
 void SnakeKun::setGameMode(GameMode _gm)
