+++ ConSnake/x64/Debug/ConSnake.tlog/ConSnake.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.26.28801:TargetPlatformVersion=10.0.18362.0:
+Debug|x64|C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\|
