+++ ConSnake/x64/Debug/ConSnake.vcxproj.FileListAbsolute.txt
@@ -1,0 +1,2 @@
+C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\x64\Debug\ConSnake.exe
+C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\x64\Debug\ConSnake.pdb
