+++ ConSnake/x64/Debug/NativeCodeAnalysis.read.1.tlog
@@ -1,0 +1,1 @@
+C:\PROGRAM FILES (X86)\MICROSOFT VISUAL STUDIO\2019\COMMUNITY\TEAM TOOLS\STATIC ANALYSIS TOOLS\\RULE SETS\NATIVERECOMMENDEDRULES.RULESET
