+++ ConSnake/x64/Debug/main.nativecodeanalysis.xml
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<DEFECTS>
+  <DEFECT>
+    <SFA>
+      <FILEPATH>C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\</FILEPATH>
+      <FILENAME>Food.h</FILENAME>
+      <LINE>9</LINE>
+      <COLUMN>1</COLUMN>
+    </SFA>
+    <DEFECTCODE>26495</DEFECTCODE>
+    <DESCRIPTION>Variable 'Food::_coord' is uninitialized. Always initialize a member variable (type.6).</DESCRIPTION>
+    <FUNCTION>Food::{ctor}</FUNCTION>
+    <DECORATED>??0Food@@QEAA@XZ</DECORATED>
+    <FUNCLINE>9</FUNCLINE>
+    <ADDITIONALINFO>
+      <targetSymbol>Food::_coord</targetSymbol>
+    </ADDITIONALINFO>
+    <PATH></PATH>
+  </DEFECT>
+  <DEFECT>
+    <SFA>
+      <FILEPATH>C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\</FILEPATH>
+      <FILENAME>Food.cpp</FILENAME>
+      <LINE>10</LINE>
+      <COLUMN>34</COLUMN>
+    </SFA>
+    <DEFECTCODE>6385</DEFECTCODE>
+    <DESCRIPTION>Reading invalid data from 'map[_coord.y]':  the readable size is '88' bytes, but '312' bytes may be read.</DESCRIPTION>
+    <FUNCTION>Food::{ctor}</FUNCTION>
+    <DECORATED>??0Food@@QEAA@QEAY0BG@W4MapData@@@Z</DECORATED>
+    <FUNCLINE>3</FUNCLINE>
+    <PROBABILITY>1</PROBABILITY>
+    <RANK>4</RANK>
+    <CATEGORY>
+      <RULECATEGORY>mspft</RULECATEGORY>
+    </CATEGORY>
+    <PATH>
+      <SFA>
+        <FILEPATH>C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\</FILEPATH>
+        <FILENAME>Food.cpp</FILENAME>
+        <LINE>4</LINE>
+        <COLUMN>0</COLUMN>
+      </SFA>
+      <SFA>
+        <FILEPATH>C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\</FILEPATH>
+        <FILENAME>Food.cpp</FILENAME>
+        <LINE>5</LINE>
+        <COLUMN>6</COLUMN>
+      </SFA>
+      <SFA>
+        <FILEPATH>C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\</FILEPATH>
+        <FILENAME>Food.cpp</FILENAME>
+        <LINE>8</LINE>
+        <COLUMN>11</COLUMN>
+      </SFA>
+      <SFA>
+        <FILEPATH>C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\</FILEPATH>
+        <FILENAME>Food.cpp</FILENAME>
+        <LINE>9</LINE>
+        <COLUMN>11</COLUMN>
+      </SFA>
+      <SFA>
+        <FILEPATH>C:\Users\asus\Desktop\Git\SS004.K22.Bai.7-\ConSnake\</FILEPATH>
+        <FILENAME>Food.cpp</FILENAME>
+        <LINE>10</LINE>
+        <COLUMN>34</COLUMN>
+        <KEYEVENT>
+          <ID>1</ID>
+          <KIND>usage</KIND>
+          <IMPORTANCE>Essential</IMPORTANCE>
+          <MESSAGE>Invalid read from 'map[_coord.y][77]', (readable range is 0 to 21)</MESSAGE>
+        </KEYEVENT>
+      </SFA>
+    </PATH>
+  </DEFECT>
+</DEFECTS>
