--- Menu.h
@@ -1,13 +1,0 @@
-#pragma once
-#include "Library.h"
-
-class Menu
-{
-public:
-    Menu();
-    ~Menu();
-    void StartMenu();
-    GameMode GamemodeMenu();
-};
-
-
